apps/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
argon2-cffi==23.1.0
cryptography==42.0.5
pydantic==2.7.0
email-validator==2.1.1
pydantic-settings
phonenumbers==8.13.14
httpx==0.24.0
twilio==8.13.0